Extracting logic to a dedicated class
